refactor: apply destructuring, js modules
