+++ js/data.js
@@ -1,0 +1,202 @@
+export default [
+  {
+    text: ['`', '~', 'ё', 'Ё'], keyCode: 'Backquote',
+  },
+  {
+    text: ['1', '!', '1', '!'], keyCode: 'Digit1',
+  },
+  {
+    text: ['2', '@', '2', '"'], keyCode: 'Digit2',
+  },
+  {
+    text: ['3', '#', '3', '№'], keyCode: 'Digit3',
+  },
+  {
+    text: ['4', '$', '4', ';'], keyCode: 'Digit4',
+  },
+  {
+    text: ['5', '%', '5', '%'], keyCode: 'Digit5',
+  },
+  {
+    text: ['6', '^', '6', ':'], keyCode: 'Digit6',
+  },
+  {
+    text: ['7', '&', '7', '?'], keyCode: 'Digit7',
+  },
+  {
+    text: ['8', '*', '8', '*'], keyCode: 'Digit8',
+  },
+  {
+    text: ['9', '(', '9', '('], keyCode: 'Digit9',
+  },
+  {
+    text: ['0', ')', '0', ')'], keyCode: 'Digit0',
+  },
+  {
+    text: ['-', '_', '-', '_'], keyCode: 'Minus',
+  },
+  {
+    text: ['=', '+', '=', '+'], keyCode: 'Equal',
+  },
+  {
+    text: ['Backspace', 'Backspace', 'Backspace', 'Backspace'], keyCode: 'Backspace', color: 'keyboard__key--dark', size: 'keyboard__key--big',
+  },
+
+
+  {
+    text: ['Tab', 'Tab', 'Tab', 'Tab'], keyCode: 'Tab', color: 'keyboard__key--dark',
+  },
+  {
+    text: ['q', 'Q', 'й', 'Й'], keyCode: 'KeyQ',
+  },
+  {
+    text: ['w', 'W', 'ц', 'Ц'], keyCode: 'KeyW',
+  },
+  {
+    text: ['e', 'E', 'у', 'У'], keyCode: 'KeyE',
+  },
+  {
+    text: ['r', 'R', 'к', 'К'], keyCode: 'KeyR',
+  },
+  {
+    text: ['t', 'T', 'е', 'Е'], keyCode: 'KeyT',
+  },
+  {
+    text: ['y', 'Y', 'н', 'Н'], keyCode: 'KeyY',
+  },
+  {
+    text: ['u', 'U', 'г', 'Г'], keyCode: 'KeyU',
+  },
+  {
+    text: ['i', 'I', 'ш', 'Ш'], keyCode: 'KeyI',
+  },
+  {
+    text: ['o', 'O', 'щ', 'Щ'], keyCode: 'KeyO',
+  },
+  {
+    text: ['p', 'P', 'з', 'З'], keyCode: 'KeyP',
+  },
+  {
+    text: ['[', '{', 'х', 'Х'], keyCode: 'BracketLeft',
+  },
+  {
+    text: [']', '}', 'ъ', 'Ъ'], keyCode: 'BracketRight',
+  },
+  {
+    text: ['\\', '|', '\\', '/'], keyCode: 'Backslash',
+  },
+  {
+    text: ['Del', 'Del', 'Del', 'Del'], keyCode: 'Delete', color: 'keyboard__key--dark',
+  },
+
+
+  {
+    text: ['CapsLock', 'CapsLock', 'CapsLock', 'CapsLock'], keyCode: 'CapsLock', color: 'keyboard__key--dark', size: 'keyboard__key--big',
+  },
+  {
+    text: ['a', 'A', 'ф', 'Ф'], keyCode: 'KeyA',
+  },
+  {
+    text: ['s', 'S', 'ы', 'Ы'], keyCode: 'KeyS',
+  },
+  {
+    text: ['d', 'D', 'в', 'В'], keyCode: 'KeyD',
+  },
+  {
+    text: ['f', 'F', 'а', 'А'], keyCode: 'KeyF',
+  },
+  {
+    text: ['g', 'G', 'п', 'П'], keyCode: 'KeyG',
+  },
+  {
+    text: ['h', 'H', 'р', 'Р'], keyCode: 'KeyH',
+  },
+  {
+    text: ['j', 'J', 'о', 'О'], keyCode: 'KeyJ',
+  },
+  {
+    text: ['k', 'K', 'л', 'Л'], keyCode: 'KeyK',
+  },
+  {
+    text: ['l', 'L', 'д', 'Д'], keyCode: 'KeyL',
+  },
+  {
+    text: [';', ':', 'ж', 'Ж'], keyCode: 'Semicolon',
+  },
+  {
+    text: ['\'', '"', 'э', 'Э'], keyCode: 'Quote',
+  },
+  {
+    text: ['Enter', 'Enter', 'Enter', 'Enter'], keyCode: 'Enter', color: 'keyboard__key--dark', size: 'keyboard__key--big',
+  },
+
+
+  {
+    text: ['Shift', 'Shift', 'Shift', 'Shift'], keyCode: 'ShiftLeft', color: 'keyboard__key--dark', size: 'keyboard__key--big',
+  },
+  {
+    text: ['z', 'Z', 'я', 'Я'], keyCode: 'KeyZ',
+  },
+  {
+    text: ['x', 'X', 'ч', 'Ч'], keyCode: 'KeyX',
+  },
+  {
+    text: ['c', 'C', 'с', 'С'], keyCode: 'KeyC',
+  },
+  {
+    text: ['v', 'V', 'м', 'М'], keyCode: 'KeyV',
+  },
+  {
+    text: ['b', 'B', 'и', 'И'], keyCode: 'KeyB',
+  },
+  {
+    text: ['n', 'N', 'т', 'Т'], keyCode: 'KeyN',
+  },
+  {
+    text: ['m', 'M', 'ь', 'Ь'], keyCode: 'KeyM',
+  },
+  {
+    text: [',', '<', 'б', 'Б'], keyCode: 'Comma',
+  },
+  {
+    text: ['.', '>', 'ю', 'Ю'], keyCode: 'Period',
+  },
+  {
+    text: ['/', '?', '.', ','], keyCode: 'Slash',
+  },
+  {
+    text: ['↑', '↑', '↑', '↑'], keyCode: 'ArrowUp', color: 'keyboard__key--dark',
+  },
+  {
+    text: ['Shift', 'Shift', 'Shift', 'Shift'], keyCode: 'ShiftRight', color: 'keyboard__key--dark', size: 'keyboard__key--big',
+  },
+
+
+  {
+    text: ['Ctrl', 'Ctrl', 'Ctrl', 'Ctrl'], keyCode: 'ControlLeft', color: 'keyboard__key--dark',
+  },
+  {
+    text: ['Win', 'Win', 'Win', 'Win'], keyCode: 'MetaLeft', color: 'keyboard__key--dark',
+  },
+  {
+    text: ['Alt', 'Alt', 'Alt', 'Alt'], keyCode: 'AltLeft', color: 'keyboard__key--dark',
+  },
+  {
+    text: [' ', ' ', ' ', ' '], keyCode: 'Space', color: undefined, size: 'keyboard__key--biggest',
+  },
+  {
+    text: ['Alt', 'Alt', 'Alt', 'Alt'], keyCode: 'AltRight', color: 'keyboard__key--dark',
+  },
+  {
+    text: ['←', '←', '←', '←'], keyCode: 'ArrowLeft', color: 'keyboard__key--dark',
+  },
+  {
+    text: ['↓', '↓', '↓', '↓'], keyCode: 'ArrowDown', color: 'keyboard__key--dark',
+  },
+  {
+    text: ['→', '→', '→', '→'], keyCode: 'ArrowRight', color: 'keyboard__key--dark',
+  },
+  {
+    text: ['Ctrl', 'Ctrl', 'Ctrl', 'Ctrl'], keyCode: 'ControlRight', color: 'keyboard__key--dark',
+  },
+];
